requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
pycocotools>=2.0.6
scikit-learn>=1.3.0
matplotlib>=3.7.0
diffusers>=0.21.0
accelerate>=0.23.0
safetensors>=0.4.0
flask>=3.0.0
groq>=0.4.0
nltk>=3.8.0
rouge-score>=0.1.2
google-generativeai>=0.3.0
